+++ sydtest/output-test/Spec.hs
@@ -9,10 +9,10 @@ import Control.Concurrent
 import Control.Exception
 import Control.Monad
 import Data.ByteString (ByteString)
-import qualified Data.ByteString.Builder as SBB
-import qualified Data.ByteString.Lazy as LB
 import Data.List
 import Data.Text (Text)
+import qualified Data.Text.Lazy as LT
+import qualified Data.Text.Lazy.Builder as TLB
 import System.Exit
 import System.Random
 import Test.QuickCheck
@@ -144,9 +144,9 @@ spec = do
 
     describe "outputResultForest" $ do
       it "outputs the same as last time" $ do
-        pureGoldenByteStringFile
+        pureGoldenTextFile
           "test_resources/output.golden"
-          (LB.toStrict $ SBB.toLazyByteString $ renderResultReport defaultSettings With24BitColours (Timed [] 0))
+          (LT.toStrict $ TLB.toLazyText $ renderResultReport defaultSettings With24BitColours (Timed [] 0))
 
   doNotRandomiseExecutionOrder $
     describe "Around" $